+++ __tests__/utils.js
@@ -140,6 +140,60 @@ const snsEvent = {
   ]
 }
 
+// Sample event from https://docs.aws.amazon.com/lambda/latest/dg/with-sqs.html
+const sqsEvent = {
+  Records: [
+    {
+      messageId: '059f36b4-87a3-44ab-83d2-661975830a7d',
+      receiptHandle: 'AQEBwJnKyrHigUMZj6rYigCgxlaS3SLy0a...',
+      body: 'Test message.',
+      attributes: {
+        ApproximateReceiveCount: '1',
+        SentTimestamp: '1545082649183',
+        SenderId: 'AIDAIENQZJOLO23YVJ4VO',
+        ApproximateFirstReceiveTimestamp: '1545082649185'
+      },
+      messageAttributes: {},
+      md5OfBody: 'e4e68fb7bd0e697a0ae8f1bb342846b3',
+      eventSource: 'aws:sqs',
+      eventSourceARN: 'arn:aws:sqs:us-east-2:123456789012:my-queue',
+      awsRegion: 'us-east-2'
+    }
+  ]
+}
+
+// Sample event from https://docs.aws.amazon.com/lambda/latest/dg/services-cloudwatchevents.html
+const eventbridgeEvent = {
+  version: '0',
+  id: 'fe8d3c65-xmpl-c5c3-2c87-81584709a377',
+  'detail-type': 'RDS DB Instance Event',
+  source: 'aws.rds',
+  account: '123456789012',
+  time: '2020-04-28T07:20:20Z',
+  region: 'us-east-2',
+  resources: ['arn:aws:rds:us-east-2:123456789012:db:rdz6xmpliljlb1'],
+  detail: {
+    EventCategories: ['backup'],
+    SourceType: 'DB_INSTANCE',
+    SourceArn: 'arn:aws:rds:us-east-2:123456789012:db:rdz6xmpliljlb1',
+    Date: '2020-04-28T07:20:20.112Z',
+    Message: 'Finished DB Instance backup',
+    SourceIdentifier: 'rdz6xmpliljlb1'
+  }
+}
+
+const eventbridgeScheduledEvent = {
+  version: '0',
+  account: '123456789012',
+  region: 'us-east-2',
+  detail: {},
+  'detail-type': 'Scheduled Event',
+  source: 'aws.events',
+  time: '2019-03-01T01:23:45Z',
+  id: 'cdc73f9d-aea9-11e3-9d5a-835b769c0d9c',
+  resources: ['arn:aws:events:us-east-2:123456789012:rule/my-schedule']
+}
+
 describe('getEventSourceNameBasedOnEvent', () => {
   test('throws error on empty event', () => {
     expect(() => getEventSourceNameBasedOnEvent({ event: {} })).toThrow(
@@ -161,10 +215,28 @@ describe('getEventSourceNameBasedOnEvent', () => {
     const result = getEventSourceNameBasedOnEvent({ event: snsEvent })
     expect(result).toEqual('AWS_SNS')
   })
+
+  test('recognizes sqs event', () => {
+    const result = getEventSourceNameBasedOnEvent({ event: sqsEvent })
+    expect(result).toEqual('AWS_SQS')
+  })
+
+  test('recognizes eventbridge event', () => {
+    const result = getEventSourceNameBasedOnEvent({ event: eventbridgeEvent })
+    expect(result).toEqual('AWS_EVENTBRIDGE')
+  })
+
+  test('recognizes eventbridge scheduled event', () => {
+    const result = getEventSourceNameBasedOnEvent({ event: eventbridgeScheduledEvent })
+    expect(result).toEqual('AWS_EVENTBRIDGE')
+  })
 })
 
 module.exports = {
   samHttpApiEvent,
   dynamoDbEvent,
-  snsEvent
+  snsEvent,
+  sqsEvent,
+  eventbridgeEvent,
+  eventbridgeScheduledEvent
 }