+++ src/event-sources/index.js
@@ -3,8 +3,10 @@ const awsApiGatewayV2EventSource = require('./aws/api-gateway-v2')
 const awsAlbEventSource = require('./aws/alb')
 const awsLambdaEdgeEventSource = require('./aws/lambda-edge')
 const awsSnsEventSource = require('./aws/sns')
+const awsSqsEventSource = require('./aws/sqs')
 const awsDynamoDbEventSource = require('./aws/dynamodb')
 const azureHttpFunctionV3EventSource = require('./azure/http-function-runtime-v3')
+const awsEventBridgeEventSource = require('./aws/eventbridge')
 
 function getEventSource ({ eventSourceName }) {
   switch (eventSourceName) {
@@ -22,6 +24,10 @@ function getEventSource ({ eventSourceName }) {
       return awsSnsEventSource
     case 'AZURE_HTTP_FUNCTION_V3':
       return azureHttpFunctionV3EventSource
+    case 'AWS_SQS':
+      return awsSqsEventSource
+    case 'AWS_EVENTBRIDGE':
+      return awsEventBridgeEventSource
     default:
       throw new Error('Couldn\'t detect valid event source.')
   }