+++ src/event-sources/utils.js
@@ -79,6 +79,9 @@ function getEventSourceNameBasedOnEvent ({
     if (eventSource === 'aws:dynamodb') {
       return 'AWS_DYNAMODB'
     }
+    if (eventSource === 'aws:sqs') {
+      return 'AWS_SQS'
+    }
     return 'AWS_LAMBDA_EDGE'
   }
   if (event.requestContext) {
@@ -94,6 +97,26 @@ function getEventSourceNameBasedOnEvent ({
     }
 
     return 'AZURE_HTTP_FUNCTION_V3'
+  }
+  if (
+    event.version &&
+    event.version === '0' &&
+    event.id &&
+    event['detail-type'] &&
+    event.source &&
+    event.source.startsWith('aws.') && // Might need to adjust this for "Partner Sources", e.g. Auth0, Datadog, etc
+    event.account &&
+    event.time &&
+    event.region &&
+    event.resources &&
+    Array.isArray(event.resources) &&
+    event.detail &&
+    typeof event.detail === 'object' &&
+    !Array.isArray(event.detail)
+  ) {
+    // AWS doesn't have a defining Event Source here, so we're being incredibly selective on the structure
+    // Ref: https://docs.aws.amazon.com/lambda/latest/dg/services-cloudwatchevents.html
+    return 'AWS_EVENTBRIDGE'
   }
 
   throw new Error('Unable to determine event source based on event.')